Step 21: refactor keyboard input
